added a basic implementation of the testResultRepository and the TestResultRange
git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/caretaker/trunk@18276 735d13b6-9817-0410-8766-e36946ffe9aa

+++ trunk/ext_tables.sql
@@ -138,7 +138,7 @@ CREATE TABLE tx_caretaker_testresults (
 	group_uid int(11) DEFAULT '0' NOT NULL,
 	
 	result_status int(11) DEFAULT '0' NOT NULL,
-	result_value int(11) DEFAULT '0' NOT NULL,
+	result_value float DEFAULT '0' NOT NULL,
 	result_msg varchar(255) DEFAULT '' NOT NULL,
 	result_data tinytext NOT NULL,
